requirements.txt
# Dependencies
requests>=2.25.0
rich>=13.7 ; extra == "pretty"
flask>=3.1
cryptography>=45.0

# Optional: For environment variable management
python-dotenv>=0.19.0
click>=5.0

